refactor types declarations
