+++ src/messages/constants/validation.ts
@@ -7,7 +7,7 @@ const m = (function () {
   const MIN_PRICE = 1;
   const V_USER = "user";
   const MIN_GENDER = 1;
-  const MIN_STRING = 20;
+  const MIN_STRING = 2;
   const MAX_GENDER = 20;
   const MAX_STRING = 255;
   const V_EMAIL = "email";
