+++ src/types/email/index.ts
@@ -1,0 +1,16 @@
+export interface EmailOTP {
+  email: string;
+  otp: string;
+}
+
+export interface ObjectEmailAndPortType {
+  emailHost: string;
+  emailPort: number;
+}
+
+export interface ObjectEmailBody {
+  emailSentTo: string;
+  emailSubject: string;
+  emailText: string;
+  emailComposed: string;
+}
