+++ src/types/misc/index.ts
@@ -1,0 +1,10 @@
+export interface Response {
+  message: string;
+  data: object;
+  http: number;
+}
+
+export interface KeySearchObject {
+  _id: string;
+  otherField: string;
+}
