+++ src/types/otp/index.ts
@@ -1,0 +1,9 @@
+export interface OTPData {
+  _id: string;
+  otpInput: string;
+  expiresAt: number;
+}
+
+export interface ResendOTP {
+  _id: string;
+}
