+++ src/types/user/index.ts
@@ -1,0 +1,14 @@
+export interface User {
+  _id: string;
+  name: string;
+  address: string;
+  dob: Date;
+  email: string;
+  cellNumber: string;
+  gender: string;
+  password: string;
+  isVerified: boolean;
+  isResendCode: boolean;
+  otp: string;
+  expiresAt: number;
+}
